Vue Side

+++ resources/assets/js/components/posts/Posts.vue
@@ -1,94 +1,119 @@
 <template>
-  <div>
-  <div v-if="isloading">
-        <loading></loading>
-  </div>
-  <div v-if="!isloading">
-    <my-errors :msg="msg" ></my-errors>
+<div>
+    <div v-if="isloading">
+        <loadinger :message="message"></loadinger>
+    </div>
+    <div v-if="!isloading">
     <div v-if="isall">
-      <create-post :posts="posts"></create-post>
-      <h1>All Posts</h1>
+      <create-post @eventname="updateparent" :posts.sync="posts"></create-post>
+      <h1 class="text-danger">All Posts</h1>
+       <nav aria-label="Navigation">
+            <ul class="pagination justify-content-end">
+                <li class="page-item" 
+                v-bind:class="[{disabled: !pagination.prev_page_url}]">
+                <a class="page-link" @click="allPosts(pagination.prev_page_url)"><ic icon="backward" size="1x"></ic></a>
+                </li>
+
+                <li class="page-item disabled">
+                <a class="page-link text-secondary font-weight-bold">Page {{pagination.current_page}} of {{pagination.last_page}}</a>
+                </li>
+
+                <li class="page-item" v-bind:class="[{disabled: !pagination.next_page_url}]">
+                <a class="page-link" @click="allPosts(pagination.next_page_url)"><ic icon="forward" size="1x"></ic></a>
+                </li>
+            </ul>
+        </nav>
       <div v-for="post in posts" v-bind:key="post.id">
-        <div class="jumbotron p-2">
-          <button @click="singlePost(post.id)" type="button" class="btn btn-info col-12">View</button>
-          <h2>{{post.title}}</h2>
-          <small>Written On {{post.created_at}} By {{post.user.first}} {{post.user.last}}</small>
-        </div>
+            <div class="jumbotron p-2">
+                <button @click="singlePost(post.id)" class="btn btn-primary col-12 mb-2"><ic icon="eye" size="lg" class="text-light animated infinite bounceIn"></ic> <ic icon="eye" size="lg" class="text-light ml-2 animated infinite bounceIn"></ic></button>
+                <h3 class="text-center text-danger">{{post.title}}</h3>
+                <p class="help-block text-center text-warning">Written On {{post.created_at}} By {{post.user.first}} {{post.user.last}}</p>
+            </div>
       </div>
     </div>
     <div v-if="!isall">
+      <button @click="allPosts()" class="btn btn-light col-12 mt-2 mb-2"><ic icon="users"></ic> <b>Posts</b></button>
       <div v-if="isedit" class="jumbotron p-2">
       <edit-post v-if="access" :post=post></edit-post>
-      <button v-if="access" @click="deletePost(post.id)" class="btn btn-primary col-12 mt-2">Delete</button>
-      <button @click="willEdit(false)" type="button" class="btn btn-secondary col-12 mt-2">Cancel</button>
+      <button v-if="access" @click="deletePost(post.id)" class="btn btn-info col-12 mt-2 text-light"><ic icon="trash-alt"></ic> <b>Delete</b></button>
+      <button @click="willEdit(false)" class="btn btn-dark col-12 mt-2"><ic icon="backward" size="lg"></ic> <b>Back</b></button>
       </div>
-      <div class="jumbotron p-2">
+        <div v-if="isreport" class="jumbotron p-2">
+        <report-user :post="post"></report-user>
+      <button @click="willReport(false)" class="btn btn-dark col-12 mt-2"><ic icon="backward"></ic> <b>Back</b></button>
+        </div>
         <div v-if="access">
          <div v-if="!isedit">
-            <button @click="willEdit(true)" type="button" class="btn btn-danger col-12 mt-2">Edit your post</button>
+            <button @click="willEdit(true)" type="button" class="btn btn-success col-12 mb-3 text-light"><ic icon="edit" size="lg"></ic> <b>Edit your post</b></button>
           </div>
         </div> 
-        <button @click="allPosts" type="button" class="btn btn-secondary col-12 mt-2">Back</button>
-        <h3>{{post.title}}</h3>
-        <small>Written On {{post.created_at}} By {{post.user.first}} {{post.user.last}}</small>
-        <p>{{post.body}}</p>
-          <button @click="reactPost(post.id,true)" class="btn badge badge-danger">Like {{post.react.like}}</button>
-          <button @click="reactPost(post.id,false)" class="btn badge badge-dark">Dislike {{post.react.dislike}}</button>
-        <h5><b>Comments</b></h5>
-        <div v-for="comment in comments" v-bind:key="comment.id">
-          <p>{{comment.user_id.first}} {{comment.user_id.last}}</p>
-          <small>Written On {{comment.created_at}}</small>
-          <p>{{comment.body}}</p>
-          <span class="btn badge badge-danger">Like {{comment.react.like}}</span>
-          <span class="btn badge badge-dark">Dislike {{comment.react.dislike}}</span>
-          <hr>
+        <div v-if="!access">
+           <div v-if="!isreport">
+            <button @click="willReport(true)" type="button" class="btn btn-warning col-12 mb-3 text-light"><ic icon="exclamation-triangle" size="lg"></ic> <b>Report</b></button>
+          </div>
+        </div> 
+        <div class="jumbotron p-2 text-danger">
+        <div class="jumbotron p-2 bg-dark">
+        <h2>{{post.title}}</h2>
+        <p class="help-block text-warning h6">Written On {{post.created_at}} By {{post.user.first}} {{post.user.last}}</p>
+          <loading :active.sync="wait" :can-cancel="true"></loading>
+          <h5>{{post.body}}</h5>
+           <ic @click="reactPost(post.id,true)" icon="thumbs-up" size="lg"></ic> <b>{{post.react.like}}</b>
+          <ic @click="reactPost(post.id,false)" icon="thumbs-down" class="text-info ml-2" size="lg"></ic> <span class="text-info"><b>{{post.react.dislike}}</b></span>
         </div>
+        <h5><b>Comments</b></h5>
+        <create-comment :post="post"></create-comment>
       </div>
     </div>
   </div>
 </div>
 </template>
 <script>
-import Handle_Message from '../errors/Handle_Message.vue'
-import Loading from '../inc/Loading.vue'
+import Loadings from '../inc/Loading.vue'
 import CreatePost from './CreatePost.vue'
 import EditPost from './EditPost.vue'
+import swal from 'sweetalert'
+import CreateComment from '../comments/CreateComment.vue'
+import ReportUser from '../reports/ReportUser.vue'
+import Loading from 'vue-loading-overlay';
+import 'vue-loading-overlay/dist/vue-loading.min.css';
   export default{
     data(){
       return {
         posts:[],
         post:[],
-        comments: [],
         access: false,
         isall: true,
-        msg :{
-            iserror: false,
-            issuccess: false,
-            message: null
-        },
         isloading: false,
         isedit: false,
         data : {
           react : null
         },
+        message:{
+          title: "Post"
+        },
+        isreport: false,
+        pagination: {},
+        wait: false
       }
     },
     components: {
-        'my-errors' : Handle_Message,
-        'loading': Loading,
+        'report-user': ReportUser,
+        'loadinger': Loadings,
         'create-post': CreatePost,
         'edit-post': EditPost,
+        'create-comment': CreateComment,
+        Loading
     },
     created(){
       this.allPosts()
+      this.data.react = 'like'
+      console.log(this.data.react)
     },
     methods: {
       clearData(all){
-        if(all)
-          this.posts = []
-        else
-          this.post = []
-          this.comments = []
+          this.isreport = false
+          this.isedit = false
       },
       willEdit(edit){
         if(edit)
@@ -96,17 +121,19 @@ import EditPost from './EditPost.vue'
         else
          this.isedit = false
       },
-      allPosts(){
+      allPosts(page_url){
         var vm = this
         this.isloading = true
         this.isedit = false
-        this.$http.get('api/posts',{
+        page_url = page_url || 'api/posts';
+        this.$http.get(page_url,{
           headers: {
                  Authorization: 'Bearer ' + this.$auth.getToken()
           }
         })
         .then(function(response){
-          vm.posts = response.data
+          vm.posts = response.data.data
+          vm.makePagination(response.data.meta,response.data.links);
           this.isall = true
           this.clearData(false)
           this.isloading = false
@@ -115,6 +142,15 @@ import EditPost from './EditPost.vue'
           console.log(error)
           vm.isloading = false
         })
+      },
+      makePagination(meta,links) {
+          let pagination = {
+              current_page: meta.current_page,
+              last_page: meta.last_page,
+              next_page_url: links.next,
+              prev_page_url: links.prev
+          }
+          this.pagination = pagination;
       },
       accessControl(id){
         var vm = this
@@ -137,7 +173,6 @@ import EditPost from './EditPost.vue'
         var vm = this
         vm.accessControl(id)
         this.isloading = true
-        vm.allComments(id)
         this.$http.get('api/posts/'+id,{
           headers: {
                  Authorization: 'Bearer ' + this.$auth.getToken()
@@ -156,60 +191,68 @@ import EditPost from './EditPost.vue'
       },
       deletePost(id){
         var vm = this
-        vm.isloading = true
-        this.$http.delete('api/posts/'+id,{
-          headers: {
-                 Authorization: 'Bearer ' + this.$auth.getToken()
+          swal({
+            title: "Are you sure ??",
+            text: "Do you want to delete your post ?", 
+            icon: "warning",
+            buttons: true,
+            dangerMode: true,
+        })
+        .then((willDelete) => {
+          if (willDelete) {
+            swal("Your post is delete!", {
+              icon: "success",
+            });
+             this.$http.delete('api/posts/'+id,{
+              headers: {
+                    Authorization: 'Bearer ' + this.$auth.getToken()
+              }
+            })
+            .then(function(response){
+           //   console.log(response.data)
+              this.allPosts()
+            })
+            .catch(function(error){
+              console.log(error)
+            })
+          } else {
+            swal("Your decision changed");
+                
           }
         })
-        .then(function(response){
-          console.log(response.data)
-          this.allPosts()
-        })
-        .catch(function(error){
-          console.log(error)
-          vm.isloading = false
-        })
-      },
-      allComments(id){
-        var vm = this
-        this.$http.get('api/posts/'+ id +'/comments',{
-          headers: {
-                 Authorization: 'Bearer ' + this.$auth.getToken()
-          }
-        })
-        .then(function(response){
-          if(response.data.empty){
-              vm.comments = []
-              return false
-          }
-          vm.comments = response.data
-        })
-        .catch(function(error){
-          console.log(error)
-
-        })
+       
       },
       reactPost(id,islike){
         var vm = this
-        this.data.react = 'like'
-        if(!islike)
+        this.wait = true
+        if(!islike) {
           this.data.react = 'dislike'
-       
-        this.$http.post('api/react/posts/'+ id,vm.data,{
+        } else {
+          this.data.react = 'like'
+            islike = false
+        }
+        this.$http.post('api/react/posts/'+ id,this.data,{
           headers: {
                  Authorization: 'Bearer ' + this.$auth.getToken()
           }
         })
         .then(function(response){
           vm.post.react = response.data.react
+          vm.wait = false
         })
         .catch(function(error){
-          vm.post.react = response.data.react
           console.log(error)
         })
-    
-    }
+      },
+      updateparent(variable) {
+        this.posts = variable
+      },
+      willReport(report){
+        if(report)
+         this.isreport = true
+        else
+         this.isreport = false
+      },
   }
 }
 </script>
